Move AutoComplete into types/project.
