--- src/types/AutoComplete.tsx
@@ -1,4 +1,0 @@
-export enum AutoComplete {
-  Off = "Off",
-  On = "On",
-}
